requirements.txt
streamlit>=1.24.0
torch>=1.10.0
torchvision>=0.11.0
faiss-cpu>=1.7.3
pillow>=9.0.0
numpy>=1.22.0
pandas>=1.4.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
tqdm>=4.64.0
open-clip-torch>=2.0.0
transformers>=4.20.0
huggingface-hub>=0.10.0